Dpkg::Shlibs: Don't resolve library directory symlinks
This breaks on Endless where /lib is a symlink to /usr/lib but the
binary packages still use the non-symlinked path. E.g., it would prefer
/usr/lib/ld-linux.so.2 even though libc6 stores the path as
/lib/ld-linux.so.2.

The case this was trying to handle is the opposite of what we're trying
to do. The resolved path (/usr/lib) is the canonical path for the
original (/usr/lib64). For us, the original is the canonical path.

[endlessm/eos-shell#5276]