Dpkg::Vendor::Ubuntu: Set Xdg-App binary control field
If the xdg-app build profile is in use, look for appdata files in
/app/share/appdata. If a desktop type appdata file is found, set the
Xdg-App control field to the value in the <id> subelement with the
.desktop suffix stripped. This will allow packages representing xdg-apps
to be found from the repository by looking for the Xdg-App control
field.

The Xdg-App field is also registered, so a package can override the
setting in the control file by applying it to the package containing the
application.

This should probably be in an Endless specific vendor module, but until
there's a larger audit of packages, we're using Ubuntu as our vendor.
The XML::LibXML module is required rather than used to avoid a hard
runtime dependency. This is to hopefully allow perl ABI upgrades if we
ever need that.